Remove unnecessary changes

+++ pandas/core/arrays/string_.py
@@ -290,7 +290,7 @@ class StringArray(PandasArray):
     def value_counts(self, dropna=False):
         from pandas import value_counts
 
-        return value_counts(self._ndarray, dropna=dropna).astype("Int64")
+        return value_counts(self._ndarray, dropna=dropna)
 
     # Override parent because we have different return types.
     @classmethod
