+++ pandas/tests/indexes/period/test_ops.py
@@ -47,7 +47,7 @@ class TestPeriodIndexOps:
             ],
             freq="H",
         )
-        expected = Series(range(10, 0, -1), index=exp_idx, dtype="Int64")
+        expected = Series(range(10, 0, -1), index=exp_idx, dtype="int64")
 
         for obj in [idx, Series(idx)]:
             tm.assert_series_equal(obj.value_counts(), expected)
@@ -68,13 +68,13 @@ class TestPeriodIndexOps:
         )
 
         exp_idx = PeriodIndex(["2013-01-01 09:00", "2013-01-01 08:00"], freq="H")
-        expected = Series([3, 2], index=exp_idx, dtype="Int64")
+        expected = Series([3, 2], index=exp_idx)
 
         for obj in [idx, Series(idx)]:
             tm.assert_series_equal(obj.value_counts(), expected)
 
         exp_idx = PeriodIndex(["2013-01-01 09:00", "2013-01-01 08:00", NaT], freq="H")
-        expected = Series([3, 2, 1], index=exp_idx, dtype="Int64")
+        expected = Series([3, 2, 1], index=exp_idx)
 
         for obj in [idx, Series(idx)]:
             tm.assert_series_equal(obj.value_counts(dropna=False), expected)
