+++ pandas/tests/indexes/timedeltas/test_ops.py
@@ -18,7 +18,7 @@ class TestTimedeltaIndexOps:
         idx = TimedeltaIndex(np.repeat(idx.values, range(1, len(idx) + 1)))
 
         exp_idx = timedelta_range("1 days 18:00:00", freq="-1H", periods=10)
-        expected = Series(range(10, 0, -1), index=exp_idx, dtype="Int64")
+        expected = Series(range(10, 0, -1), index=exp_idx, dtype="int64")
 
         for obj in [idx, Series(idx)]:
             tm.assert_series_equal(obj.value_counts(), expected)
@@ -38,13 +38,13 @@ class TestTimedeltaIndexOps:
         )
 
         exp_idx = TimedeltaIndex(["1 days 09:00:00", "1 days 08:00:00"])
-        expected = Series([3, 2], index=exp_idx, dtype="Int64")
+        expected = Series([3, 2], index=exp_idx)
 
         for obj in [idx, Series(idx)]:
             tm.assert_series_equal(obj.value_counts(), expected)
 
         exp_idx = TimedeltaIndex(["1 days 09:00:00", "1 days 08:00:00", pd.NaT])
-        expected = Series([3, 2, 1], index=exp_idx, dtype="Int64")
+        expected = Series([3, 2, 1], index=exp_idx)
 
         for obj in [idx, Series(idx)]:
             tm.assert_series_equal(obj.value_counts(dropna=False), expected)
